task delete impl

+++ src/main/java/ozzy/project/todoApp/db/entity/task/TaskDAOImpl.java
@@ -3,7 +3,11 @@ package ozzy.project.todoApp.db.entity.task;
 import org.springframework.stereotype.Repository;
 
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityTransaction;
 import jakarta.persistence.PersistenceContext;
+import jakarta.transaction.Transactional;
+import ozzy.project.todoApp.db.entity.user.User;
+import ozzy.project.todoApp.db.entity.user.UserRepository;
 
 @Repository
 public class TaskDAOImpl implements TaskDAO {
@@ -11,13 +15,25 @@ public class TaskDAOImpl implements TaskDAO {
     @PersistenceContext
     private final EntityManager entityManager;
 
-    public TaskDAOImpl(EntityManager entityManager) {
+    private final UserRepository userRepository;
+
+    public TaskDAOImpl(EntityManager entityManager, UserRepository userRepository) {
         this.entityManager = entityManager;
+        this.userRepository = userRepository;
     }
 
     @Override
-    public void delete(Task task) {
-        entityManager.remove(task);
+    @Transactional
+    public void delete(Integer id) {
+        Task task = entityManager.find(Task.class, id);
+        if (task != null) {
+            User user = task.getUser();
+            if (user != null) {
+                user.getTasks().remove(task);
+                userRepository.save(user);
+            }
+            entityManager.remove(task);
+        }
     }
 
     @Override
